NeighborlyAPI/requirements.txt
adal==1.2.7
aniso8601==9.0.1
applicationinsights==0.11.10
argcomplete==1.12.3
attrs==20.3.0
azure-cli-core==2.22.1
azure-cli-telemetry==1.0.6
azure-common==1.1.27
azure-core==1.13.0
azure-cosmos==4.2.0
azure-functions==1.7.0
azure-mgmt-core==1.2.2
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.2
dominate==2.6.0
feedgen==0.9.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
flask-restplus==0.13.0
flask-swagger-ui==3.36.0
humanfriendly==9.1
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
knack==0.8.1
lxml==4.6.3
MarkupSafe==1.1.1
msal==1.11.0
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.1.0
paramiko==2.7.2
pkginfo==1.7.0
portalocker==1.7.1
psutil==5.8.0
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pymongo==3.11.3
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tabulate==0.8.9
urllib3==1.26.4
visitor==0.1.3
Werkzeug==0.16.1
